Set back float64 to float32.

+++ lib/improver/nowcasting/optical_flow.py
@@ -202,7 +202,7 @@ class OpticalFlow(object):
                 np.diff(data, axis=axis), axis=1-axis)
             outdata.append(diffs)
         smoothed_diffs = np.zeros(
-            [self.shape[0]+1, self.shape[1]+1], dtype=np.float64)
+            [self.shape[0]+1, self.shape[1]+1], dtype=np.float32)
         smoothed_diffs[1:-1, 1:-1] = 0.5*(outdata[0] + outdata[1])
         return self.interp_to_midpoint(smoothed_diffs)
 
@@ -219,7 +219,7 @@ class OpticalFlow(object):
         """
         tdiff = self.data2 - self.data1
         smoothed_diffs = np.zeros(
-            [self.shape[0]+1, self.shape[1]+1], dtype=np.float64)
+            [self.shape[0]+1, self.shape[1]+1], dtype=np.float32)
         smoothed_diffs[1:-1, 1:-1] = self.interp_to_midpoint(tdiff)
         return self.interp_to_midpoint(smoothed_diffs)
 
@@ -254,7 +254,7 @@ class OpticalFlow(object):
                     (self.data2[i:i+self.boxsize, j:j+self.boxsize]).sum())
                 weight = 1. - np.exp(-1.*weight/0.8)
                 weights.append(weight)
-        weights = np.array(weights, dtype=np.float64)
+        weights = np.array(weights, dtype=np.float32)
         weights[weights < 0.01] = 0
         return boxes, weights
 
@@ -274,7 +274,7 @@ class OpticalFlow(object):
         grid_data = np.repeat(np.repeat(box_data, self.boxsize, axis=0),
                               self.boxsize, axis=1)
         grid_data = grid_data[:self.shape[0],
-                              :self.shape[1]].astype(np.float64)
+                              :self.shape[1]].astype(np.float32)
         return grid_data
 
     @staticmethod
@@ -365,7 +365,7 @@ class OpticalFlow(object):
         # define kernel for neighbour weighting
         neighbour_kernel = (np.array([[0.5, 1, 0.5],
                                       [1.0, 0, 1.0],
-                                      [0.5, 1, 0.5]])/6.).astype(np.float64)
+                                      [0.5, 1, 0.5]])/6.).astype(np.float32)
 
         # smooth input data and weights fields
         vel_neighbour = scipy.ndimage.convolve(weights*vel_iter,
@@ -532,8 +532,8 @@ class OpticalFlow(object):
         # (c) Reshape displacement arrays to match array of subbox points
         newshape = [int((self.shape[0]-1)/self.boxsize) + 1,
                     int((self.shape[1]-1)/self.boxsize) + 1]
-        umat = np.array(velocity[0], dtype=np.float64).reshape(newshape)
-        vmat = np.array(velocity[1], dtype=np.float64).reshape(newshape)
+        umat = np.array(velocity[0], dtype=np.float32).reshape(newshape)
+        vmat = np.array(velocity[1], dtype=np.float32).reshape(newshape)
         weights = box_weights.reshape(newshape)
 
         # (d) Check for extreme advection displacements (over a significant
@@ -724,7 +724,7 @@ class OpticalFlow(object):
         # calculate smoothing radius in grid square units
         new_coord = cube1.coord(axis='x').copy()
         new_coord.convert_units('km')
-        grid_length_km = np.float64(np.diff((new_coord).points)[0])
+        grid_length_km = np.float32(np.diff((new_coord).points)[0])
         data_smoothing_radius = \
             int(data_smoothing_radius_km / grid_length_km)
 
@@ -761,15 +761,15 @@ class OpticalFlow(object):
             msg = ("No non-zero data in input fields: setting optical flow "
                    "velocities to zero")
             warnings.warn(msg)
-            ucomp = np.zeros(data1.shape, dtype=np.float64)
-            vcomp = np.zeros(data2.shape, dtype=np.float64)
+            ucomp = np.zeros(data1.shape, dtype=np.float32)
+            vcomp = np.zeros(data2.shape, dtype=np.float32)
         else:
             # calculate dimensionless displacement between the two input fields
             ucomp, vcomp = self.process_dimensionless(data1, data2, 1, 0,
                                                       data_smoothing_radius)
             # convert displacements to velocities in metres per second
             for vel in [ucomp, vcomp]:
-                vel *= np.float64(1000.*grid_length_km)
+                vel *= np.float32(1000.*grid_length_km)
                 vel /= cube_time_diff.total_seconds()
 
         # create velocity output cubes based on metadata from later input cube
